requirements.txt
google-generativeai          0.7.2
streamlit                    1.38.0